pyproject.toml
[project]
name = "notification-system"
version = "0.1.0"
description = "Notification system"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "psycopg[binary] (>=3.3.2,<4.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "alembic (>=1.18.1,<2.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
package-mode = false

